Pull boolean synopsis out of unit tests and use it
So far this synopsis was just an implementation detail in the meta
index. Now it's enabled for all bool types by default.
