+++ libvast/CMakeLists.txt
@@ -15,6 +15,7 @@ set(libvast_sources
   src/banner.cpp
   src/base.cpp
   src/bitmap.cpp
+  src/boolean_synopsis.cpp
   src/chunk.cpp
   src/column_index.cpp
   src/column_major_matrix_table_slice_builder.cpp
